+++ src/test/java/domain/result/BettingResultTest.java
@@ -1,9 +1,15 @@
-package domain.game;
+package domain.result;
 
+import domain.TestDeck;
 import domain.card.Card;
 import domain.card.CardNumber;
 import domain.card.CardType;
 import domain.card.Deck;
+import domain.gamer.Dealer;
+import domain.gamer.Gamer;
+import domain.gamer.Name;
+import domain.gamer.Player;
+import domain.gamer.Players;
 import domain.money.Money;
 import java.util.List;
 import java.util.Map;
